move internals of blake2 to the core package.
